standard test impl names: "integration" => "standard"
