+++ airbyte-integrations/connectors/destination-postgres/src/test-integration/java/io/airbyte/integrations/destination/postgres/PostgresStandardTest.java
@@ -22,7 +22,7 @@
  * SOFTWARE.
  */
 
-package io.airbyte.integrations.destination.mysql;
+package io.airbyte.integrations.destination.postgres;
 
 import com.fasterxml.jackson.databind.JsonNode;
 import com.google.common.collect.ImmutableMap;
@@ -30,31 +30,25 @@ import io.airbyte.commons.json.Jsons;
 import io.airbyte.db.Databases;
 import io.airbyte.integrations.base.JavaBaseConstants;
 import io.airbyte.integrations.destination.ExtendedNameTransformer;
-import io.airbyte.integrations.standardtest.destination.TestDestination;
+import io.airbyte.integrations.standardtest.destination.DestinationStandardTest;
 import java.sql.SQLException;
+import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Collectors;
 import org.jooq.JSONFormat;
 import org.jooq.JSONFormat.RecordFormat;
-import org.jooq.SQLDialect;
-import org.junit.jupiter.api.Test;
-import org.testcontainers.containers.MySQLContainer;
+import org.testcontainers.containers.PostgreSQLContainer;
 
-public class MySQLIntegrationTest extends TestDestination {
+public class PostgresStandardTest extends DestinationStandardTest {
 
   private static final JSONFormat JSON_FORMAT = new JSONFormat().recordFormat(RecordFormat.OBJECT);
 
-  private MySQLContainer<?> db;
-  private ExtendedNameTransformer namingResolver = new MySQLNameTransformer();
+  private PostgreSQLContainer<?> db;
+  private ExtendedNameTransformer namingResolver = new ExtendedNameTransformer();
 
   @Override
   protected String getImageName() {
-    return "airbyte/destination-mysql:dev";
-  }
-
-  @Override
-  protected boolean implementsNamespaces() {
-    return true;
+    return "airbyte/destination-postgres:dev";
   }
 
   @Override
@@ -63,8 +57,10 @@ public class MySQLIntegrationTest extends TestDestination {
         .put("host", db.getHost())
         .put("username", db.getUsername())
         .put("password", db.getPassword())
-        .put("database", db.getDatabaseName())
+        .put("schema", "public")
         .put("port", db.getFirstMappedPort())
+        .put("database", db.getDatabaseName())
+        .put("ssl", false)
         .build());
   }
 
@@ -74,43 +70,62 @@ public class MySQLIntegrationTest extends TestDestination {
         .put("host", db.getHost())
         .put("username", db.getUsername())
         .put("password", "wrong password")
-        .put("database", db.getDatabaseName())
+        .put("schema", "public")
         .put("port", db.getFirstMappedPort())
+        .put("database", db.getDatabaseName())
+        .put("ssl", false)
         .build());
   }
 
   @Override
-  protected String getDefaultSchema(JsonNode config) {
-    if (config.get("database") == null) {
-      return null;
-    }
-    return config.get("database").asText();
-  }
-
-  @Override
-  protected List<JsonNode> retrieveRecords(TestDestinationEnv testEnv,
-                                           String streamName,
-                                           String namespace)
-      throws Exception {
+  protected List<JsonNode> retrieveRecords(TestDestinationEnv env, String streamName, String namespace) throws Exception {
     return retrieveRecordsFromTable(namingResolver.getRawTableName(streamName), namespace)
         .stream()
         .map(r -> Jsons.deserialize(r.get(JavaBaseConstants.COLUMN_NAME_DATA).asText()))
         .collect(Collectors.toList());
   }
 
+  @Override
+  protected boolean implementsBasicNormalization() {
+    return true;
+  }
+
+  @Override
+  protected boolean implementsNamespaces() {
+    return true;
+  }
+
+  @Override
+  protected List<JsonNode> retrieveNormalizedRecords(TestDestinationEnv env, String streamName, String namespace)
+      throws Exception {
+    String tableName = namingResolver.getIdentifier(streamName);
+    // Temporarily disabling the behavior of the ExtendedNameTransformer, see (issue #1785) so we don't
+    // use quoted names
+    // if (!tableName.startsWith("\"")) {
+    // // Currently, Normalization always quote tables identifiers
+    // //tableName = "\"" + tableName + "\"";
+    // }
+    return retrieveRecordsFromTable(tableName, namespace);
+  }
+
+  @Override
+  protected List<String> resolveIdentifier(String identifier) {
+    final List<String> result = new ArrayList<>();
+    final String resolved = namingResolver.getIdentifier(identifier);
+    result.add(identifier);
+    result.add(resolved);
+    if (!resolved.startsWith("\"")) {
+      result.add(resolved.toLowerCase());
+      result.add(resolved.toUpperCase());
+    }
+    return result;
+  }
+
   private List<JsonNode> retrieveRecordsFromTable(String tableName, String schemaName) throws SQLException {
-    return Databases.createDatabase(
-        db.getUsername(),
-        db.getPassword(),
-        String.format("jdbc:mysql://%s:%s/%s",
-            db.getHost(),
-            db.getFirstMappedPort(),
-            db.getDatabaseName()),
-        "com.mysql.cj.jdbc.Driver",
-        SQLDialect.MYSQL).query(
+    return Databases.createPostgresDatabase(db.getUsername(), db.getPassword(),
+        db.getJdbcUrl()).query(
             ctx -> ctx
-                .fetch(String.format("SELECT * FROM %s.%s ORDER BY %s ASC;", schemaName, tableName,
-                    JavaBaseConstants.COLUMN_NAME_EMITTED_AT))
+                .fetch(String.format("SELECT * FROM %s.%s ORDER BY %s ASC;", schemaName, tableName, JavaBaseConstants.COLUMN_NAME_EMITTED_AT))
                 .stream()
                 .map(r -> r.formatJSON(JSON_FORMAT))
                 .map(Jsons::deserialize)
@@ -119,53 +134,14 @@ public class MySQLIntegrationTest extends TestDestination {
 
   @Override
   protected void setup(TestDestinationEnv testEnv) {
-    db = new MySQLContainer<>("mysql:8.0");
+    db = new PostgreSQLContainer<>("postgres:13-alpine");
     db.start();
-    setLocalInFileToTrue();
-    revokeAllPermissions();
-    grantCorrectPermissions();
-  }
-
-  private void setLocalInFileToTrue() {
-    executeQuery("set global local_infile=true");
-  }
-
-  private void revokeAllPermissions() {
-    executeQuery("REVOKE ALL PRIVILEGES, GRANT OPTION FROM " + db.getUsername() + "@'%';");
-  }
-
-  private void grantCorrectPermissions() {
-    executeQuery("GRANT CREATE, INSERT, SELECT, DROP ON *.* TO " + db.getUsername() + "@'%';");
-  }
-
-  private void executeQuery(String query) {
-    try {
-      Databases.createDatabase(
-          "root",
-          "test",
-          String.format("jdbc:mysql://%s:%s/%s",
-              db.getHost(),
-              db.getFirstMappedPort(),
-              db.getDatabaseName()),
-          "com.mysql.cj.jdbc.Driver",
-          SQLDialect.MYSQL).query(
-              ctx -> ctx
-                  .execute(query));
-    } catch (SQLException e) {
-      throw new RuntimeException(e);
-    }
   }
 
   @Override
   protected void tearDown(TestDestinationEnv testEnv) {
     db.stop();
     db.close();
-  }
-
-  @Override
-  @Test
-  public void testLineBreakCharacters() {
-    // overrides test with a no-op until we handle full UTF-8 in the destination
   }
 
 }
